Refactoring methods into stand alone function classes
